requirements.txt
absl-py==1.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aioice==0.10.1
aiortc==1.11.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
argcomplete==3.5.3
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==23.1.0
av==14.3.0
backoff==2.2.1
backports.cached-property==1.0.2
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.36.13
botocore==1.36.13
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.1.0
cryptography==44.0.2
cycler==0.11.0
dacite==1.7.0
datasets==3.3.2
debugpy==1.8.12
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
durationpy==0.9
eventlet==0.39.0
executing==2.2.0
fastapi==0.115.12
fiftyone==0.16.3
fiftyone-brain==0.8.2
fiftyone-db==0.3.0
filelock==3.17.0
Flask==3.1.0
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
glob2==0.7
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-crc32c==1.7.1
google-pasta==0.2.0
googleapis-common-protos==1.69.2
graphql-core==3.1.7
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h2==4.2.0
h5py==3.9.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.1
humanfriendly==10.0
humanize==4.11.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.4
ifaddr==0.2.0
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflate64==1.0.1
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.0.5
kaleido==0.2.1
keras==2.14.0
kiwisolver==1.4.4
kubernetes==32.0.1
lazy_loader==0.4
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.3
ml-dtypes==0.2.0
mmh3==5.1.0
mongoengine==0.20.0
monotonic==1.6
motor==2.5.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multivolumefile==0.2.3
narwhals==1.25.1
ndjson==0.3.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.21.0
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opencv-python-headless==4.11.0.86
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
opt-einsum==3.3.0
orjson==3.10.16
overrides==7.7.0
packaging==23.1
pandas==2.2.3
parso==0.8.4
patool==3.1.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
posthog==3.23.0
pprintpp==0.4.0
priority==2.0.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==4.25.6
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
py4j==0.10.9.7
py7zr==0.22.0
pyarrow==19.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.3
pycparser==2.21
pycryptodomex==3.21.0
pydantic==2.11.0
pydantic_core==2.33.0
pydash==8.0.5
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
pylibsrtp==0.12.0
pymongo==3.13.0
pyOpenSSL==25.0.0
pyparsing==3.0.9
PyPika==0.48.9
pyppmd==1.1.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyspark==3.4.4
python-dateutil==2.8.2
python-dotenv==1.1.0
python-multipart==0.0.5
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
pyzstd==0.16.2
rarfile==4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
s3transfer==0.11.2
safetensors==0.5.2
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentence-transformers==4.0.1
sentinel==0.3.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.6
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.46.1
strawberry-graphql==0.96.0
streamlit==1.42.0
streamlit-option-menu==0.4.0
streamlit-webrtc==0.62.4
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2025.1.10
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
ultralytics==8.3.111
ultralytics-thop==2.0.14
universal-analytics-python3==1.1.1
urllib3==1.26.16
uvicorn==0.34.0
voxel51-eta==0.7.1
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
